pyproject.toml
[project]
name = "echo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Laboratorio IA" }
]
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "duckdb>=1.4.4",
    "fastapi[standard]>=0.128.0",
    "livekit-agents>=1.3.12",
    "opentelemetry-exporter-otlp>=1.39.1",
    "pydantic>=2.12.5",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.14",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = false
disallow_untyped_decorators = false
disallow_subclassing_any = false
strict = true
exclude = ['^setup\.py$', "venv"]

[[tool.mypy.overrides]]
module = [
    "pymupdf.*",
    "docx.*",
    "fitz.*",
    "lxml.*",
    "langgraph.*",
    "langchain_core.*",
    "langchain_openai.*",
]

[[tool.mypy.overrides]]
module = [
]
ignore_errors = true

[tool.ruff]
lint.select = [
    "E",   # pycodestyle
    "W",   # pycodestyle
    "F",   # pyflake
    "I",   # isort
    "S",   # flake8-bandit
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "D",   # pydocstyle
    "RUF", # Ruff specific rules
]

lint.ignore = [
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D106",    # Missing docstring in public nested class
    "D105",    # Missing docstring in magic method
    "D107",    # Missing docstring in __init__
    "D203",    # no-blank-line-before-class (conflict with D211)
    "D213",    # multi-line-summary-second-line (conflict with D212)
    "S314",    # use defusedxml instead of xml
    "S320",    # Using `lxml` to parse untrusted data is known to be vulnerable to XML attacks
    "S608",    # SQL Injection - Should be enabled at some point?
    "UP022",   # Prefer `capture_output` over sending `stdout` and `stderr` to `PIPE`
    "S603",    # `subprocess` call: check for execution of untrusted input
    "RUF001",  # Unambiguous unicode character
]

exclude = [
    ".git",
    ".idea",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "build",
    "assets/*/*",
    "docs/",
]

line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104"]
"tests/*" = ["S", "D"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.coverage.run]
source = ["src", "tests"]

[tool.hatch.metadata]
allow-direct-references = true

